kinship: lift shared AI-REML loop into reml.rs, split sparse into folder
Restructure of the kinship module so contributors can read the math in
one place. No algorithm change. Statistical invariance verified: 161/161
tests pass, including the dense bit-identity check
(assemble_sigma_single_kinship_single_group), the sparse-vs-dense parity
test, and the seeded-Hutchinson reproducibility test.

Layout

  src/staar/kinship/
    mod.rs           dispatcher: warm-start, path selection, fit_reml
    types.rs         (unchanged)
    load.rs          (unchanged)
    budget.rs        (unchanged)
    pql.rs           (unchanged)
    reml.rs          NEW. SigmaSolver enum, SolverBuilder trait, ai_step,
                     run_reml. The shared AI-REML algorithm and convergence
                     loop. This is the file to read.
    dense.rs         REWRITTEN. DenseBuilder + assemble_sigma + dense
                     trace_p_k + fit_reml_dense entry. 510 -> 182 LoC.
    sparse/
      mod.rs         NEW. SparseBuilder + sparse trace_p_k + fit_reml_sparse
                     entry.
      assembler.rs   NEW. Assembler (was SparseSigmaAssembler) — cached
                     union pattern + symbolic Cholesky.
      hutchinson.rs  NEW. trace_k_estimate (was hutchinson_trace),
                     diag_inverse_estimate (was hutchinson_diag_inverse),
                     sparse_matvec, xorshift PRNG. Marked as the fallback
                     until the Takahashi recursion in #26 #27 lands.
    UPSTREAM.md      NEW. Pinned GMMAT and GENESIS SHAs, function map,
                     constant provenance, variable rename map.

Abstraction

The dense and sparse paths previously had two near-copies of the AI step
(dense.rs::ai_step, sparse.rs::ai_step_sparse) and two copies of the
convergence loop (fit_reml_inner, fit_reml_inner_sparse). Both now go
through one ai_step and one converge() in reml.rs.

The trick is two small abstractions:

  trait SolverBuilder { fn build(&self, tau) -> Result<SigmaSolver>; }

builds a fresh SigmaSolver each AI iteration. DenseBuilder materializes
n*n Sigma_inv; SparseBuilder factors via the cached Assembler.

  enum SigmaSolver { Dense(Mat), Sparse(SparseSolverState) }

provides Sigma_inv * v, diag(Sigma_inv), and tr(P K_l) at the rate the
shared ai_step needs them. The trace term dispatches to a per-backend
free function (dense::trace_p_k_dense, sparse::trace_p_k_sparse) so each
path keeps its original floating-point arithmetic order — that's how
statistical invariance is preserved without unifying the trace formulas.

Renames

  SparseSigmaAssembler   ->  Assembler
  EntryContrib           ->  PatternEntry
  hutchinson_trace       ->  trace_k_estimate
  hutchinson_diag_inverse ->  diag_inverse_estimate
  apply_p                ->  apply_projection (now a free function in reml.rs)
  weights_or_one         ->  weights_or_ones
  ai_step_sparse / fit_reml_inner_sparse → gone, replaced by the shared
                                            ai_step / converge.

Upstream pinning + citations

UPSTREAM.md captures the bits Phase 0 of this audit dug up:

  hanchenphd/GMMAT@473b34268ade7f8b920f67657c6f4759880d35d3 (2025-11-21)
  UW-GAC/GENESIS@dd480380c7705d36d1dbeb09c73b08f60623701a   (2026-03-04)

Every math function in the new layout has a `// upstream:` comment at
its header pointing at the right line range. Three things came out of
that walkthrough that the old comments got wrong or omitted:

1. The old dense.rs cited R_fitglmm_ai but that's the *sparse* path in
   upstream R (R/glmmkin.R:662). The dense R path is R_fitglmm_ai_dense
   (R/glmmkin.R:712). Citations now point at the correct functions.

2. R_fitglmm_ai_var doesn't exist in upstream R — the old sparse.rs
   citation was made up. Sparse path is now correctly cited against
   R_fitglmm_ai (R/glmmkin.R:662-710).

3. Of the five constants in the inner loop, three are upstream
   (REML_TOL=1e-5, REML_MAX_ITER=500, BOUNDARY_FACTOR=1.01) and two are
   ours as safety nets (REML_MAX_OUTER_REFITS=10, REML_STEP_HALVE_MAX=50).
   Upstream's corresponding while loops are unbounded; we cap and return
   an Analysis error on hit. Documented in UPSTREAM.md "constant
   provenance".

GENESIS validates the two-loop PQL structure we use in pql.rs:
.iterateAIREMLworkingY in R/iterateAIREMLworkingY.R:17-63 has the same
nested shape (outer loop updates working response, inner loop runs full
AI-REML). GMMAT interleaves the working-response update inside the AI
loop. Both converge to the same point.

Sparse path is still not bit-identical to upstream

This is unchanged from before the refactor — flagging it because the
audit makes it explicit. Upstream R_fitglmm_ai computes the trace term
sum(Sigma_i * kins[[i]]) exactly via dense Sigma_i (chol2inv on a sparse
Cholesky materializes the full inverse). We keep Sigma_i unmaterialized
and use a Hutchinson estimator (~3% relative error at M=30 probes).
Phase 4 (Takahashi recursion, #26 #27) replaces Hutchinson with an exact
computation of Sigma_i entries restricted to the union sparsity pattern
of the kinship matrices, which is exactly what the Frobenius product
needs. After that lands the sparse path is 1:1 with upstream while
still using strictly less memory.

LoC

Pre-refactor: dense.rs 510 + sparse.rs 1072 = 1582 LoC.
Post-refactor: reml.rs 712 + dense.rs 182 + sparse/{mod.rs 387,
assembler.rs 229, hutchinson.rs 151} = 1661 LoC. Net +79 LoC. Not the
"reduce before you add" win the user usually wants — the trait machinery
and the doc-comment citations both add bytes. The dense path code
collapses 67% (485 -> 157 LoC excluding tests); the sparse path code
collapses ~30% (820 -> 570 LoC excluding tests); reml.rs is brand new
infrastructure (+712 LoC). Trade is accepted because the math is now in
one readable place and every formula has its upstream citation.

Tests

161/161 pass. Kinship subset:

  staar::kinship::dense::tests::assemble_sigma_single_kinship_single_group
  staar::kinship::sparse::tests::sparse_assembler_pattern_includes_diagonal
  staar::kinship::sparse::tests::sparse_factor_succeeds_at_warm_start
  staar::kinship::sparse::tests::hutchinson_trace_reproducible
  staar::kinship::sparse::tests::sparse_fit_reml_pedigree_matches_dense_within_tolerance
  staar::kinship::sparse::tests::sparse_fit_reml_reproducible
  staar::kinship::tests::fit_reml_groups_only_matches_ols_sigma2
  staar::kinship::tests::fit_reml_three_groups_recovers_per_group_variance
  staar::kinship::tests::fit_reml_kinship_block_family_converges
  staar::kinship::tests::fit_pql_glmm_runs_and_returns_state
  + 10 in types.rs / budget.rs

The sparse-vs-dense parity test is the load-bearing one — it confirms
Phase 1A and 1B did not change either backend's arithmetic.
